+++ tests/test_database.py
@@ -27,10 +27,19 @@ async def test_create_drop_database(url, sys_db_name, root, password):
         sys_db = await client.db(
             sys_db_name, auth_method="basic", auth=auth, verify=True
         )
+
+        # Create a new database
         db_name = generate_db_name()
         assert await sys_db.create_database(db_name) is True
         await client.db(db_name, auth_method="basic", auth=auth, verify=True)
         assert await sys_db.has_database(db_name) is True
+
+        # List available databases
+        dbs = await sys_db.databases()
+        assert db_name in dbs
+        assert "_system" in dbs
+
+        # Drop the newly created database
         assert await sys_db.delete_database(db_name) is True
         non_existent_db = generate_db_name()
         assert await sys_db.has_database(non_existent_db) is False
@@ -46,10 +55,16 @@ async def test_create_drop_collection(url, sys_db_name, root, password):
     # TODO also handle exceptions
     async with ArangoClient(hosts=url) as client:
         db = await client.db(sys_db_name, auth_method="basic", auth=auth, verify=True)
+
+        # Create a new collection
         col_name = generate_col_name()
         col = await db.create_collection(col_name)
         assert isinstance(col, StandardCollection)
         assert await db.has_collection(col_name)
+        cols = await db.collections()
+        assert any(c.name == col_name for c in cols)
+
+        # Drop the newly created collection
         assert await db.delete_collection(col_name) is True
         assert not await db.has_collection(col_name)
         non_existent_col = generate_col_name()
